requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
python-docx
kaleido